+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+- Use asyncio event loop for send_telemetry
 
 ## [0.6.5] - 2022-04-18
 - Upgrade and freeze pyright version